Parsing input
The input from the  .lib file is parsed as the function name, the number of parameters, and the delay of the function. The .cir file initializes the map of string and boolean and initializes the inputs to 0, then the components' name is extracted for two reasons yet to be implemented. First, to call the function; second, to know the number of parameters; to know the delay of the gate, for example, if we have AND2, AND3.